# ddtrace/tracer: Remove log.Debug from hot path (-2 allocs; -5% CPU)