--- a/internal/log/log.go
+++ b/internal/log/log.go
@@ -56,6 +56,15 @@ func SetLevel(lvl Level) {
 	level = lvl
 }
 
+// DebugEnabled returns true if debug log messages are enabled. This can be used in extremely
+// hot code paths to avoid allocating the ...interface{} argument.
+func DebugEnabled() bool {
+	mu.RLock()
+	lvl := level
+	mu.RUnlock()
+	return lvl == LevelDebug
+}
+
 // Debug prints the given message if the level is LevelDebug.
 func Debug(fmt string, a ...interface{}) {
 	mu.RLock()